requirements.txt
# Used by pip to install required python packages
# Usage: pip install -r requirements.txt
# Project implements MySQL, Install MySQL Server 8.0.32 in your environment


chatterbot==1.0.4
chatterbot-corpus>=1.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.23
protobuf==3.15.5
psycopg2==2.9.3
pycparser==2.20
PyMySQL==1.0.2
python-dateutil<=2.8
python-editor==1.0.4
six==1.15.0
SQLAlchemy<1.3
Werkzeug==1.0.1
WTForms==2.3.3